+++ src/Areas/Grafana/Options/Workspace/WorkspaceListOptions.cs
@@ -1,0 +1,8 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+using AzureMcp.Options;
+
+namespace AzureMcp.Areas.Grafana.Options.Workspace;
+
+public class WorkspaceListOptions : SubscriptionOptions;