+++ src/Program.cs
@@ -63,6 +63,7 @@ internal class Program
             new AzureMcp.Areas.AzureIsv.AzureIsvSetup(),
             new AzureMcp.Areas.Cosmos.CosmosSetup(),
             new AzureMcp.Areas.Foundry.FoundrySetup(),
+            new AzureMcp.Areas.Grafana.GrafanaSetup(),
             new AzureMcp.Areas.KeyVault.KeyVaultSetup(),
             new AzureMcp.Areas.Kusto.KustoSetup(),
             new AzureMcp.Areas.Monitor.MonitorSetup(),