+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.3.2 (Unreleased)
 
 ### Features Added
+- Support for Azure Managed Grafana operations - List Azure Managed Grafana details. [#532](https://github.com/Azure/azure-mcp/pull/532)
 
 ### Breaking Changes